struct creation binding
Some APIs use struct values for options etc. To create a "MyStruct", you can now call "my-struct" and set any fields you need to.

Unfortunately, to make this work, it was necessary to make all structs pass-by-reference (i.e. turn them into pointers). This is because Rye Native values don't inherently have an address associated, making it impossible to set a struct's field without making it a pointer or creating a copy (I've opted for pointer, because it's less confusing to the user).