Fix to underlying conversion (somewhat)
The issue is that a go type may be typedefed to a numeric type e.g. type MyInt int. In that case, MyInt should still be able to do math operations like *,+ etc. However, MyInt may also have methods attached to it, which we would lose if we automatically converted to an int.

Although this commit doesn't fully address this dilemma, it now converts to an underlying type only if the typedefed type has no methods attached to it.

This, for example, (re-)fixes Fyne's NewList, while also leaving go-imap binding gen functional.