Redesign converter, auto-track conv dependencies
This will hopefully make the converter a lot easier to use (although I still have to update main.go and ryegen_test.go, but it's late so not now).

Generally, the idea is to offload most of the work to converter.ConverterSet.

Now the converter tracks dependencies and imports completely automatically. This may unlock the ability in the future to have the user add custom converters without much hassle or the need to recompile Ryegen.

I also changed the tests to include their converter dependencies.